cogl-gst: Add a function to facilitate layering
This functions is used when attaching frames at a start
point other than layer 0 is required. This could potentially
be used to "layer" videos and textures on top of each other
in the CoglPipeline. This layering could come handy if videos
are used as alpha masks or normal maps, or when arranging
layers in a perticular order, so Cogl could blend them
nicely without extra hassle.